✨ feat: Add form category and splitting algorithm selection in Create Project page
This commit adds the functionality to select form category and splitting algorithm in the Create Project page. The form category and splitting algorithm data are fetched from API using CreateProjectService and FormCategoryService respectively. The selected values are stored in Redux store using CreateProjectActions. The components used for selection are InputLabel and Select from MUI. If splitting algorithm is 'Divide on Square', a text field for dimension is also displayed.